+++ HISTORY.rst
@@ -7,3 +7,17 @@ History
 ==========
 
 - Skeleton with basic tacticals for implicational propositional logic.
+
+2016-02-24
+==========
+ 
+- Added succ_tac, fail_tac, or_tac, and try_tac
+- Tactics now are expected to return goal trees. Goal tree
+  constructors were renamed to make their use more intuitive.
+
+
+2016-02-25
+==========
+
+- Added then_tac, thenL_tac, and repeat_tac
+- Added more proofs in proofs.ml